# openapi.yaml
openapi: 3.1.0
info:
  title: Preloop AI API
  description: REST API for Preloop AI issue tracking management
  version: 0.1.0
paths:
...
  /api/v1/notification-preferences/me:
    get:
      tags:
      - Notification Preferences
      summary: Get My Notification Preferences
      description: "Get current user's notification preferences.\n\nReturns:\n   \
        \ Notification preferences."
      operationId: get_my_notification_preferences_api_v1_notification_preferences_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
      security:
      - bearerAuth: []
    put:
      tags:
      - Notification Preferences
      summary: Update My Notification Preferences
      description: "Update current user's notification preferences.\n\nArgs:\n   \
        \ prefs_in: Preferences update data.\n\nReturns:\n    Updated notification\
        \ preferences."
      operationId: update_my_notification_preferences_api_v1_notification_preferences_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferencesUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /api/v1/notification-preferences/me/register-device:
    post:
      tags:
      - Notification Preferences
      summary: Register Mobile Device
      description: "Register a mobile device for push notifications.\n\nArgs:\n  \
        \  device_in: Device registration data.\n\nReturns:\n    Updated notification\
        \ preferences."
      operationId: register_mobile_device_api_v1_notification_preferences_me_register_device_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileDeviceRegistration'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /api/v1/notification-preferences/me/device/{token}:
    delete:
      tags:
      - Notification Preferences
      summary: Unregister Mobile Device
      description: "Unregister a mobile device and delete associated API key.\n\n\
        Args:\n    token: Device token to remove.\n\nReturns:\n    Updated notification\
        \ preferences.\n\nRaises:\n    HTTPException: If preferences not found."
      operationId: unregister_mobile_device_api_v1_notification_preferences_me_device__token__delete
      security:
      - bearerAuth: []
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notification-preferences/me/qr-code:
    get:
      tags:
      - Notification Preferences
      summary: Get Registration Qr Code
      description: "Generate a QR code for mobile device registration.\n\nReturns:\n\
        \    QR code data with registration token."
      operationId: get_registration_qr_code_api_v1_notification_preferences_me_qr_code_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCodeResponse'
      security:
      - bearerAuth: []
  /api/v1/notification-preferences/register-via-token:
    post:
      tags:
      - Notification Preferences
      summary: Register Device Via Token
      description: "Register a mobile device using a QR code token and create an API\
        \ key.\n\nThis endpoint is called by the mobile app after scanning the QR\
        \ code.\nIt registers the device for push notifications AND creates an API\
        \ key\nfor the mobile app to authenticate API requests.\n\nArgs:\n    token:\
        \ Registration token from QR code.\n    device_in: Device registration data.\n\
        \nReturns:\n    Updated notification preferences with API key.\n\nRaises:\n\
        \    HTTPException: If token is invalid or expired."
      operationId: register_device_via_token_api_v1_notification_preferences_register_via_token_post
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
          title: Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileDeviceRegistration'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileDeviceRegistrationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MobileDeviceRegistration:
      properties:
        platform:
          type: string
          title: Platform
          description: 'Device platform: ''ios'' or ''android'''
        token:
          type: string
          title: Token
          description: Device push notification token
        device_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Name
          description: Optional device name for API key
      type: object
      required:
      - platform
      - token
      title: MobileDeviceRegistration
      description: Schema for registering a mobile device.
    MobileDeviceRegistrationResponse:
      properties:
        preferences:
          $ref: '#/components/schemas/NotificationPreferencesResponse'
        api_key:
          type: string
          title: Api Key
          description: API key for mobile app authentication
        api_key_id:
          type: string
          format: uuid
          title: Api Key Id
          description: API key ID
        api_key_expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Api Key Expires At
          description: API key expiration
      type: object
      required:
      - preferences
      - api_key
      - api_key_id
      title: MobileDeviceRegistrationResponse
      description: Schema for mobile device registration response with API key.
    NotificationPreferencesResponse:
      properties:
        preferred_channel:
          type: string
          title: Preferred Channel
          description: 'Preferred notification channel: ''email'' or ''mobile_push'''
          default: email
        enable_email:
          type: boolean
          title: Enable Email
          description: Whether email notifications are enabled
          default: true
        enable_mobile_push:
          type: boolean
          title: Enable Mobile Push
          description: Whether mobile push notifications are enabled
          default: false
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        mobile_device_tokens:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Mobile Device Tokens
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - created_at
      - updated_at
      title: NotificationPreferencesResponse
      description: Schema for notification preferences response.
    NotificationPreferencesUpdate:
      properties:
        preferred_channel:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Channel
          description: Preferred notification channel
        enable_email:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Email
          description: Enable email notifications
        enable_mobile_push:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enable Mobile Push
          description: Enable mobile push notifications
      type: object
      title: NotificationPreferencesUpdate
      description: Schema for updating notification preferences.
...
    QRCodeResponse:
      properties:
        token:
          type: string
          title: Token
          description: Registration token
        qr_data:
          type: string
          title: Qr Data
          description: QR code data (URL)
        expires_at:
          type: string
          title: Expires At
          description: Token expiry timestamp
        expires_in_seconds:
          type: integer
          title: Expires In Seconds
          description: Seconds until expiry
      type: object
      required:
      - token
      - qr_data
      - expires_at
      - expires_in_seconds
      title: QRCodeResponse
      description: Schema for QR code registration response.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication